refactor: move parse_filter to lib.rs
